Move class CompilationController declaration to header   update impacted cpp files accordingly
